[ refactor ] cleanup yesterday's work
