+++ src/Shonkier.hs
@@ -1,8 +1,5 @@
 module Shonkier where
 
-import Control.Monad.Trans
-import qualified Data.Map as Map
-
 import Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text.IO as TIO
@@ -11,27 +8,29 @@ import Data.Text.Prettyprint.Doc
 import Data.Text.Prettyprint.Doc.Render.Text
 
 import Shonkier.Pretty ()
+import Shonkier.Value
+import Shonkier.Import
 import Shonkier.Semantics
 import Shonkier.ShonkierJS
 import Shonkier.Syntax
 
-onShonkierModule :: (Term -> ShonkierT IO a)
+onShonkierModule :: (Module -> GlobalEnv -> Term -> IO a)
                  -> String -> IO a
-onShonkierModule action filename = flip evalShonkierT emptyShonkierState $ do
-  mod@(is, ps) <- forceImportModule filename
+onShonkierModule action filename = do
+  (mod@(is, ps), gl) <- importToplevelModule filename
   case [ m | ("main", Right m) <- ps ] of
-    [([],body)] -> action body
+    [([],body)] -> action mod gl body
     _ -> error "not exactly one main function"
 
 interpretShonkier :: String -> IO ()
-interpretShonkier = onShonkierModule $ \ body ->
-  eval (Map.empty, body) >>= \case
-    Value v -> liftIO $ putDoc $ pretty v <> line
+interpretShonkier = onShonkierModule $ \ _ gl body ->
+  case shonkier gl body of
+    Value v -> putDoc $ pretty v <> line
     Request (r,_) fr -> error $ "unhandled request " ++ r
 
 -- no support for imports here yet!
 compileShonkier :: String -> IO Text
-compileShonkier = onShonkierModule $ \ body -> liftIO $ do
+compileShonkier = onShonkierModule $ \ _ _ body -> do
   -- Couldn't figure how to import in node so I just concat the
   -- whole interpreter defined 'Shonkier.js' on top
   interpreter <- TIO.readFile "./src/Shonkier/Shonkier.js"
