+++ src/Shonkier/Semantics.hs-boot
@@ -1,0 +1,6 @@
+module Shonkier.Semantics where
+
+import Shonkier.Value
+
+use :: Value -> Shonkier Computation
+handle :: Request -> [Frame] -> Shonkier Computation
