+++ src/Utils/List.hs
@@ -1,0 +1,6 @@
+module Utils.List where
+
+padCat :: Eq a => [[a]] -> [[a]] -> [[a]]
+padCat [] hs = hs
+padCat hs [] = hs
+padCat (a : as) (b : bs) = (a ++ b) : padCat as bs
